pull checkbox stories into separate folder
