+++ src/core/components/checkbox/checkbox.stories.tsx
@@ -1,0 +1,11 @@
+import { Checkbox } from "./index"
+
+export default {
+	title: "Checkbox",
+	component: Checkbox,
+}
+export * from "./stories/checkbox/indeterminate"
+export * from "./stories/checkbox/supporting-text"
+export * from "./stories/checkbox/supporting-text-only"
+export * from "./stories/checkbox/ungrouped"
+export * from "./stories/checkbox/unlabelled"
